backend/requirements.txt
fastapi
python-dotenv
uvicorn  # Recommended ASGI server for FastAPI
langchain_community 
pypdf
docx2txt
python-multipart
azure-storage-blob
